Add track command with function
This command will query an API using this guide as an example
https://blog.alexellis.io/golang-json-api-client/

It will then extrapolate the data and present the user with the current
price of Bitcoin. I found parsing data... hard, so I searched for an
abstraction to make it easier. The njson package is really nice and easy
to understand. I used this guide to help me:
https://dev.to/m7shapan/golang-how-to-unmarshal-a-subset-of-nested-json-data-d84