Replace log.Fatal with log.Panic
This commit connects to
https://github.com/jasonBirchall/crypto/issues/35 and relates to
deepsources explanation that "log.Fatal clutters `defer
res.Body.Close()`". Looking on a few forums it appears log.Fatal is the
wrong way to handle errors with a defer in the function. This
stackoverflow answer sums it up quite nicely:
https://stackoverflow.com/questions/17888610/are-deferred-functions-called-when-calling-log-fatalln

This is a nice explanation of panic:
"Panic is a built-in function that stops the ordinary flow of control
and begins panicking. When the function F calls panic, execution of F
stops, any deferred functions in F are executed normally, and then F
returns to its caller. To the caller, F then behaves like a call to
panic. The process continues up the stack until all functions in the
current goroutine have returned, at which point the program crashes.
Panics can be initiated by invoking panic directly. They can also be
caused by runtime errors, such as out-of-bounds array accesses."
Panic is a built-in function that stops the ordinary flow of control and
begins panicking. When the function F calls panic, execution of F stops,
any deferred functions in F are executed normally, and then F returns to
its caller. To the caller, F then behaves like a call to panic. The
process continues up the stack until all functions in the current
goroutine have returned, at which point the program crashes. Panics can
be initiated by invoking panic directly. They can also be caused by
runtime errors, such as out-of-bounds array accesses."
https://blog.golang.org/defer-panic-and-recover